tests/e2e/pyproject.toml
[project]
name = "e2e"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "pytest>=8.4.1",
  "pytest-sugar>=1.0.0",
  "pyyaml>=6.0.2",
  "requests>=2.32.4",
]

[dependency-groups]
dev = ["ruff>=0.12.2"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
extend-select = [
  # pep8-naming
  "N",
  # pyupgrade
  "UP",
  # isort
  "I",
  # Pyflakes
  "F",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
